Add unit test for git.commits api
Add a test to validate the git.commits api. We use the first ever commit
contributed into the spectrometer repo as our test data and validate
that values returned are as expected.

Tests will now use a unique tmpdir and their own copies of config.py and
repositories.yaml and a local clone of the spectrometer git repo to run
the tests.

Also fix pytest to only run tests once.

Change-Id: I9d0e1a9c899bfd192379f1bcf48dfb761a7a82a7
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>