Refactor initialization scripts
Summary:
- Fixes issue preventing auto reloader from reloading code in dev
- Refactor multiprocessing into a task scheduler
- Use APScheduler for task scheduling
- Refactor init scripts to support gunicorn and cli startup modes
- Fix Rest API docs linking

The Flask reloader problem was caused by the multiprocessing module
holding it's own system process which Flask's reloader couldn't reload
so was not able to automatically reload when it detected Spectrometer
code change.

Instead implement APScheduler as a background task manager to run the
database collector. This will allow us to add additional interval tasks
in the future as we need them. If we ever need a more complex scheduler
we could investigate Celery in the future.

Also refactor init scripts so that they support starting up in dev mode
via cli scripts as well as production mode using a WSGI system such as
gunicorn.

Change-Id: Ic98dc71f9db573b26916949e8ca277bac31746cd
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>