Add API to return commits since ref
/git/commits_since_ref/

    branch: branch to get commit log from
    ref: reference point to search until. ref can be any commit hash as
         well as branches and tags.

Searches Git for a common_parent between *branch* and *ref* and returns
the commit log of all the commits until the common parent excluding the
common_parent commit itself.

Also implement a utils function check_parameters(mapping) for finding
missing required parameters for the API.

Example:

    mapping = {
        'project': request.args.get('project', None),
        'branch': request.args.get('branch', 'master'),
    }

In this example 'project' is a required parameter so the get() method
defaults to None indicating that it is a required parameter. On the other
hand 'branch' is an optional parameter so we set a default value master in
the event that branch is not explicitly passed.

Change-Id: I4168e023b77bfddbb6f72057e849925ba2dffa17
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>