Cache data on demand
There are situations where we cannot predict what to cache due to it
requiring user input such as commits_since_ref api. This patch adds a
function that allows us to cache data when the api is called and checks
the cache_age before determining if we should provide the user with the
cache version or we need to update the cache.

Two new functions are introduced via this patch:

get_cache(): This function tries to query the database for our cached
    document if possible. Otherwise it will query the API for fresh data
    and calls set_cache() in order to set cached data.

set_cache(): This function is a simple function that is used to store a
    document into cache.

Also updates 3 Git APIs with the ability to cache.

    git.branches
    git.commits
    git.commits_since_ref

Change-Id: I2f62424ab9e255b50434a315eab7695ff4741803
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>