Do not sort commits by author_date
Spectrometer API returns the commits in the correct order in decending
order. Sorting by author date actually reorganizes the commits in the
incorrect order.

It is entirely possible for someone to author a patch and not get it
merged for a long time thus causing the author date to not appear in git
log chronologically.

Also has the benefit of increasing the speed of mapCommits() due to
unnecessary sorting.

Change-Id: Ia4f2756947f043aa54ba661732b13c789bcbd696
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>