Implement a simple cache via Python dictionary
Reimplement our cache system as a simple Python dictionary. Starting up
a full mongo server just to use it as a caching key-value store is
overly complicated for what we are actually using it for. Simplify the
server code by using a fast Python in-memory dictionary as the cache
system.

- Improve commit fetching speed by caching the commit.stats

Change-Id: I98bec2925694925690765185c5abcd5b5d347dcc
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>