Remove mongodb cache
With the implementation of I98bec2925694925690765185c5abcd5b5d347dcc
it has solved the root cause of our performance woes and the current
mongo cache solution seems redundant, unnecessary, and overly complex
for what we wanted to achieve with it. Scrape the mongo service for
now. We can implement a generic caching service in the future to make
it possible to swap out the backend in the future if we find that the
Python in-memory dictionary cache solution no longer serves our
needs.

Change-Id: I6735532aeab327c3a73f49502933646cc12b7b9f
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>