Store all handlers in a global dictionary
It's inefficient to be creating and destroying handlers constantly.
Since we use them a lot lets store them in a global variable so that
we can recall them easily. This patch stores the handlers into the
global dictionary githandlers as they are needed.

Move create_handler() into the git handlers file to make it easier for
other APIs to reuse this convenience function.

Change-Id: Iae1c091603341fbe1715d05ac8109054b00d8a54
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>